requirements.txt
numpy
scikit-learn
scipy
cython
xlrd
nltk
pandas
flask
gunicorn
tweet-preprocessor